fix(deprecated: code + tests)
All calls to a deprecated method or constructor have been handled in the core hierarchy. The deprecated methods remain untouched, when it was possible, the methods calling them have been turned into a version using the alternate methods provided rather than their deprecated versions.
When it was not, a @SuppressWarnings mention has been added.